Internal Documentation

Documentation for MinFEM.jl's internal types and functions.

See the Public page of the library for the documentation of the public interface.

Contents

Internal Documentation

Contents

Types

Mesh Import

Quadrature

Utility

Index

Types

MinFEM.Entity — Type

Entity

Type holding the associated physical tag and the set of elements for one gmsh elementary entity.

Mesh Import

MinFEM.import_mesh1 — Function

import_mesh1(f::IOStream) -> Mesh

Returns mesh by continuing import process started by import_mesh() for gmsh files of version 1.

MinFEM.import_mesh2 — Function

import_mesh2(f::IOStream) -> Mesh

Returns mesh by continuing import process started by import_mesh() for gmsh files of version 2.

MinFEM.import_mesh4 — Function

import_mesh4(f::IOStream) -> Mesh

Returns mesh by continuing import process started by import_mesh() for gmsh files of version 4.

MinFEM.getDimFromGMSHElementType — Function

getDimFromGMSHElementType(t::Int64) -> Int64

Returns dimension of a gmsh elementary entitiy type.

MinFEM.getGMSHElementTypeFromDim — Function

getGMSHElementTypeFromDim(d::Int64) -> Int64

Returns gmsh elementary element type for given dimension.

MinFEM.getParentBoundary — Function

getParentBoundary(nodes::Vector{Int64}, parentNodes::Vector{Int64}) -> Int64

Returns index of the boundary of the parent element spanned by the nodes.

Quadrature

MinFEM.gausslegendre_points — Function

gausslegendre_points(order::Int64) -> Array{Float64,1}

Returns coordinates of the Gauss-Legendre quadrature points on the default interval [-1,1] for exact integration of polynomials up to the given order.

MinFEM.gausslegendre_weights — Function

gausslegendre_weights(order::Int64) -> Array{Float64,1}

Returns weights of the Gauss-Legendre quadrature points on the default interval [-1,1] for exact integration of polynomials up to the given order.

MinFEM.compute_coordinates_line — Function

compute_coordinates_line(order::Int64) -> Array{Array{Float64,1},1}

Returns coordinates of the Gauss-Legendre quadrature points on the 1-dimensional FEM reference element [0,1] for exact integration of polynomials up to the given order.

MinFEM.compute_coordinates_triangle — Function

compute_coordinates_triangle(order::Int64) -> Array{Array{Float64,1},1}

Returns coordinates of the Gauss-Legendre quadrature points on the 2-dimensional FEM reference element for exact integration of polynomials up to the given order.

MinFEM.compute_coordinates_tetrahedron — Function

compute_coordinates_tetrahedron(order::Int64) -> Array{Array{Float64,1},1}

Returns coordinates of the Gauss-Legendre quadrature points on the 3-dimensional FEM reference element for exact integration of polynomials up to the given order.

MinFEM.compute_weights_line — Function

compute_weights_line(order::Int64) -> Array{Float64,1}

Returns weights of the Gauss-Legendre quadrature points on the 1-dimensional FEM reference element [0,1] for exact integration of polynomials up to the given order.

MinFEM.compute_weights_triangle — Function

compute_weights_triangle(order::Int64) -> Array{Float64,1}

Returns weights of the Gauss-Legendre quadrature points on the 2-dimensional FEM reference element for exact integration of polynomials up to the given order.

MinFEM.compute_weights_tetrahedron — Function

compute_weights_tetrahedron(order::Int64) -> Array{Float64,1}

Returns weights of the Gauss-Legendre quadrature points on the 3-dimensional FEM reference element for exact integration of polynomials up to the given order.

Utility

MinFEM.stressTensor — Function

stressTensor(grad::AbstractMatrix, lambda::Float64, mu::Float64)

Returns the local stress tensor for linear elasticity with constant coefficients λ, μ and gradient as argument.

MinFEM.strainTensor — Function

strainTensor(grad::AbstractMatrix)

Returns the local strain tensor for linear elasticity with gradient as argument.

Index

MinFEM.Entity

MinFEM.compute_coordinates_line

MinFEM.compute_coordinates_tetrahedron

MinFEM.compute_coordinates_triangle

MinFEM.compute_weights_line

MinFEM.compute_weights_tetrahedron

MinFEM.compute_weights_triangle

MinFEM.gausslegendre_points

MinFEM.gausslegendre_weights

MinFEM.getDimFromGMSHElementType

MinFEM.getGMSHElementTypeFromDim

MinFEM.getParentBoundary

MinFEM.import_mesh1

MinFEM.import_mesh2

MinFEM.import_mesh4

MinFEM.strainTensor

MinFEM.stressTensor